refactor: naver/service 패키지 구조를 service/naver 로 변경
