+++ src/main/java/com/ludo/study/studymatchingplatform/auth/controller/naver/LoginController.java
@@ -10,8 +10,8 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import com.ludo.study.studymatchingplatform.auth.common.AuthUserPayload;
 import com.ludo.study.studymatchingplatform.auth.common.provider.JwtTokenProvider;
 import com.ludo.study.studymatchingplatform.auth.naver.repository.InMemoryClientRegistrationAndProviderRepository;
-import com.ludo.study.studymatchingplatform.auth.naver.service.NaverLoginService;
-import com.ludo.study.studymatchingplatform.auth.naver.service.dto.response.LoginResponse;
+import com.ludo.study.studymatchingplatform.auth.service.naver.NaverLoginService;
+import com.ludo.study.studymatchingplatform.auth.service.naver.dto.response.LoginResponse;
 import com.ludo.study.studymatchingplatform.user.domain.Social;
 
 import jakarta.servlet.http.Cookie;
