+++ src/main/java/com/ludo/study/studymatchingplatform/auth/controller/naver/SignUpController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.ludo.study.studymatchingplatform.auth.naver.repository.InMemoryClientRegistrationAndProviderRepository;
-import com.ludo.study.studymatchingplatform.auth.naver.service.NaverSignUpService;
-import com.ludo.study.studymatchingplatform.auth.naver.service.dto.response.SignupResponse;
+import com.ludo.study.studymatchingplatform.auth.service.naver.NaverSignUpService;
+import com.ludo.study.studymatchingplatform.auth.service.naver.dto.response.SignupResponse;
 import com.ludo.study.studymatchingplatform.user.domain.Social;
 
 import lombok.RequiredArgsConstructor;
