--- src/main/java/com/ludo/study/studymatchingplatform/auth/naver/service/dto/response/SignupResponse.java
@@ -1,6 +1,0 @@
-package com.ludo.study.studymatchingplatform.auth.naver.service.dto.response;
-
-public record SignupResponse(
-		boolean ok,
-		String message) {
-}
