+++ src/main/java/com/ludo/study/studymatchingplatform/auth/service/naver/dto/response/SignupResponse.java
@@ -1,0 +1,6 @@
+package com.ludo.study.studymatchingplatform.auth.service.naver.dto.response;
+
+public record SignupResponse(
+		boolean ok,
+		String message) {
+}
